refactor logout into redux
