+++ v2/client/src/constants/actionTypes.js
@@ -37,3 +37,4 @@ export const FETCH_OVERALL_TRAINER_FEEDBACK = 'FETCH_OVERALL_TRAINER_FEEDBACK';
 export const UPDATE_VIEW_LEVEL = 'UPDATE_VIEW_LEVEL';
 export const FETCH_LOCAL_LEAD_TRAINERS_GROUP =
   'FETCH_LOCAL_LEAD_TRAINERS_GROUP';
+export const LOGOUT = 'LOGOUT';
